Add Pop Dialog successful edit
